Remove redundant type casts to double from Mat4RotY and Mat4RotZ

+++ Mat4RotY.java
@@ -16,9 +16,9 @@ public class Mat4RotY extends Mat4Identity {
 	 *            rotation angle in radians
 	 */
 	public Mat4RotY(final double alpha) {
-		mat[0][0] = (double) Math.cos(alpha);
-		mat[2][2] = (double) Math.cos(alpha);
-		mat[2][0] = (double) Math.sin(alpha);
-		mat[0][2] = (double) -Math.sin(alpha);
+		mat[0][0] = Math.cos(alpha);
+		mat[2][2] = Math.cos(alpha);
+		mat[2][0] = Math.sin(alpha);
+		mat[0][2] = -Math.sin(alpha);
 	}
 }
